Add boilerplate code to get up and running
This commit will add all of the files required to get the app up
and running locally. At this point, no work has been done on the
app's content. All of the code in this commit is just boilerplate
generated automatically by framework7 and cordova.

I wrote some useful documentation and put it in the docs directory.
Parsing the official documentation for framework7 & cordova was
challenging so I created 'how-to-start-developing.md' to document
my steps for future reference.

Things may change over time, but I think this is a good starting
point. Next steps for me are to clean up the home page, add some
menu options, update the app icon, and maybe reorganize some files.