Add Working Navigation to the left panel
I added a menu-list to the left panel. This menu-list will serve
as navigation for the application. Each list item will open a new
page in the application (except for the last item, which will close
the menu). I also added pages for the Lifting Library, Cardio
Library, Saved orkouts, Statistics, Settings, and About pages.
These pages don't have any content yet.

And for fun, I added four buttons to the Home page. In the future,
I'd like to replace the buttons with icons for each tracker but for
now this will do.

I also encountered a weird bug with the menu-list. By default the
text would be the same color as the --f7-theme-color. I fixed it by
adding a css override to app.css. I think I may be missing
something but for now this works. I will probably have to mess with
it again when I get user themes working.

Next I need to figure out how to store and load data.