Add db-adapter.js
So it seems like IndexedDB API is the way to go for data storage
based on the information I found on developer.mozilla.org.
I wasn't sure exactly how to wire everything up so I looked around
and found that davidhealey had used this approach with his
waistline app. I referenced his db-handler.js code while I was
putting my db-adapter together. It was a big help and probably
saved me hours of trial-and-error. Another win for open source!

I'm not sure if this adapter will be enough for my needs, but it's
a starting point. I'll add additional functions as necessary.