Allow user to set appearance settings
So I did a couple of things this time. I reorganized the pages and
updated routes.js to reflect that. I also decided to put the user's
app settings in localStorage. I then created settings.js to handle
applying the user's settings when the app starts. I could have just
left the code in app.js, but I'd like to keep that file as small as
possible.

Ignore everything in src/pages/home/index.f7. I'm using that page
as a sandbox to experiment with ideas.

Now for the good stuff. I updated the settings landing page to show
a list of settings the user can tweak. Right now there are only two
real options: appearance and measurement-units. I'll get to units
in the future. This commit is concerned with appearance. It took
some trial-and-error but I successfully got the darkmode toggle and
color theme selector working. On toggle, darkmode turns on or off.
And when revisiting the page, the toggle will reflect which mode
is currently applied. Same for the color theme selector. Picking
a color theme will immediately apply it and on page revisit the
current color theme will be automatically selected in the UI.

I'm actually pretty proud of how quickly I was able to get it
working. Again, not sure if it's considered 'industry standard'
in the frontend dev world but it's good enough for me.