Fix searchbar bug
While messing around I noticed that the searchbar was only working
for food items. I think it doesn't like having two searchbars on
the page at the same time. There may be a way to work around this,
but I didn't see anything in the docs so we just have one now. When
searching, it filters list items in both the food tab and meal tab.
You can see this when you toggle between tabs. It doesn't seem like
a huge deal, so I'm not going to bother with it anymore. Frontend
development is annoying. I miss the backend.