Fix Android bugs with dbAdapter & appearance settings
For fun I decided to run the app on my Android emulator, just to
make sure that things were looking okay.

I realized that app.js wasn't setup correctly. I had placed
my db init and apply user settings logic outside of the app's
on init function. This wasn't a problem while running in the browser
but was a problem on Android. I fixed this.

I also noticed that the theme color selector in appearance settings
wasn't working as expected. I realized that the problem was with
the event listener. I had set it on click when it actually needed
to be on change. Once again, my lack of frontend experience is on
full display, lol. I fixed it and updated the selector to act as
a smart-select instead of an item-input. I like the new look
better.