Add date subnav & nutrition stats to home page
I added a subnav to the home page to let the user change dates.
I also created a toolbar that will display nutrition stats for the
day. I wanted the stats toolbar to be at the bottom, but I was
having trouble getting it to stack ontop of the tabbar. After an
hour I gave up and decided that it would live at the top of the tab
page.

I also removed the white color theme from appearance settings. It
doesn't automatically change the navbar text color, and I didn't
feel like adding a custom css rule for it. Plus, anything with
white-mode makes my eyes bleed. Truth be told, I did you a favor.