Finalize Nutrition Diary UI
I added the 8 meal areas to the nutrition diary tab. With this, I
think I've got most of the UI figured out for the nutrition diary.
I will have to write more javascript to handle loading records and
displaying the relevent info on screen, but that's outside of scope
for this git branch. As far as html code is concerned, I think this
is over 90% done. Might need to make some tweaks if/when issues
come up, but I'll cross that bridge when I get there.

Next I need to work on the training diary and body diary tabs.