Add more UI functionality to home/index.f7
For the nutrition diary, I linked the track-food-items.f7 page to
each meal block's + character (I need to switch that over to an
icon soon, forgot to do it earlier). I got the track page to look
more or less how I envisioned. There's a FAB on the left for adding
new food items (it needs to be wired up to the create-food-item.f7
page) and another FAB on the right for submiting the selected
items. I'll admit, it looks a little more cramped than I thought it
would.

For the body diary, I added a FAB to the bottom right which will
ask the user to track body weight or body part sizes. Pressing on
one will open a new page to track the chosen metric. Those pages
only have a submit button. I need to get the input figured out.
track-body-weight.f7 should be trivial... the same can't be said
for track-body-part-sizes.f7. I'm still undecided about it. I want
to have some sort of graphic of the human body on that page, and
fields for number / unit entry for each part. But I have absolutely
no clue how to make something like that. Need to do more research.

Nothing changed on the training diary tab. Still up in the air
about that one.