Finalize create-exercise UI page
After working on the create-food-item page, building the
create-exercise page was trival. There is a section for metadata at
the top, a section for required exercise equipment in the middle,
and a checkbox section at the bottom for tracking which muscles
are targeted by the exercise (this should be useful later when the
user wants to see how often they work certain muscle groups).