Add on pageReinit to home page
When the user tracks their body weight, they will return to the
home page's body tab. I wanted the body tab chart to automatically
update on return so I added an event listener for pageReinit. This
function will be responsible for refreshing the page data.

So far everything looks to be working.